email delivery config and export refactored
