+++ lib/ts/recipe/emailverification/types.ts
@@ -36,7 +36,7 @@ export type TypeInput = {
 };
 
 export type TypeNormalisedInput = {
-    emailDelivery: EmailDeliveryTypeInput<TypeEmailVerificationEmailDeliveryInput>;
+    getEmailDeliveryConfig: () => EmailDeliveryTypeInput<TypeEmailVerificationEmailDeliveryInput>;
     getEmailForUserId: (userId: string, userContext: any) => Promise<string>;
     getEmailVerificationURL: (user: User, userContext: any) => Promise<string>;
     override: {
