+++ lib/ts/recipe/thirdpartyemailpassword/recipe.ts
@@ -77,10 +77,6 @@ export default class Recipe extends RecipeModule {
         super(recipeId, appInfo);
         this.config = validateAndNormaliseUserInput(this, appInfo, config);
         this.isInServerlessEnv = isInServerlessEnv;
-        this.emailDelivery =
-            ingredients.emailDelivery === undefined
-                ? new EmailDeliveryIngredient(this.config.emailDelivery)
-                : ingredients.emailDelivery;
         {
             let builder = new OverrideableBuilder(
                 RecipeImplementation(
@@ -94,6 +90,15 @@ export default class Recipe extends RecipeModule {
             let builder = new OverrideableBuilder(APIImplementation());
             this.apiImpl = builder.override(this.config.override.apis).build();
         }
+
+        /**
+         * emailDelivery will always needs to be declared after isInServerlessEnv
+         * and recipeInterfaceImpl values are set
+         */
+        this.emailDelivery =
+            ingredients.emailDelivery === undefined
+                ? new EmailDeliveryIngredient(this.config.getEmailDeliveryConfig())
+                : ingredients.emailDelivery;
 
         this.emailVerificationRecipe =
             recipes.emailVerificationInstance !== undefined
