+++ recipe/passwordless/emaildelivery/index.js
@@ -1,0 +1,6 @@
+"use strict";
+function __export(m) {
+    for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
+}
+exports.__esModule = true;
+__export(require("../../../lib/build/recipe/passwordless/emaildelivery/services"));
