+++ recipe/thirdpartyemailpassword/emaildelivery/index.js
@@ -3,4 +3,4 @@ function __export(m) {
     for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
 }
 exports.__esModule = true;
-__export(require("../../../../lib/build/recipe/thirdpartyemailpassword/emaildelivery/services"));
+__export(require("../../../lib/build/recipe/thirdpartyemailpassword/emaildelivery/services"));
