+++ recipe/thirdpartypasswordless/emaildelivery/index.d.ts
@@ -1,4 +1,4 @@
-export * from "../../../../lib/build/recipe/emailpassword/emaildelivery/services";
+export * from "../../../lib/build/recipe/thirdpartypasswordless/emaildelivery/services";
 /**
  * 'export *' does not re-export a default.
  * import NextJS from "supertokens-node/nextjs";
@@ -6,5 +6,5 @@ export * from "../../../../lib/build/recipe/emailpassword/emaildelivery/services
  * - user add "esModuleInterop": true in their tsconfig.json file
  * - we do the following change:
  */
-import * as _default from "../../../../lib/build/recipe/emailpassword/emaildelivery/services";
+import * as _default from "../../../lib/build/recipe/thirdpartypasswordless/emaildelivery/services";
 export default _default;
